It was not possible to use AWSIdentityManager to work with a fully functional AWSSignInProvider in Objective-c or Swift because:
1) AWSIdentityManager did not restart sessions from any identity providers other than Google or Facebook.
This change makes AWSIdentityManager go through an Info.plist class list to restart sessions so that developers can add AWSSignInProviders and they will function properly.

2) It was necessary to call completeLogin which was not an exposed function.
This change exposes the interface to the completeLogin function.
(I was not able to do this with a bridging header, I got problems because this framework is a static library which does not allow Swift functions.)