Add various properties and methods to AWSIdentityManager that are needed to produce a functional app.  Methods are:
currentSignInProvider - lets you get at the user pool or developer
identity pool

providerKey - lets you get the NSUserDefaults key for the session
that the AWSSignInProvider made.  This is a needed human readable
name for sign in providers

activeProviders - returns an array of all the active providers so
your app can tell which they are, how many they are, and so you
can get specific information from a provider that is not the
current sign in provider (like attributes from a user in a linked
Cognito User Pool account while logged in at Facebook).  This also
allows a simple recursive "Log out all accounts" function instead of
one login per active account.