AWSIdentityManager fix so login errors don't leave user in bad state.
Previously AWSIdentityManager will create disabled identity id's
whenever a login occurs if the App is configured to not allow merging or
multiple logins, or at least in case of login errors.
This fixes that, so that it does not create a new identity on the way to
logging in the second identity, also in cases where errors occur in login
(cancelled, wrong password) so that it does not leave the user without
an identity id.